refactor: extract ActionResult to shared models module
Move ActionResult from agent/views.py to a new openbrowser/models.py
module that depends only on pydantic. This breaks the dependency chain
where tools/service.py pulled in agent/views.py (and its heavy LLM
imports) just for ActionResult.

- Create src/openbrowser/models.py with ActionResult (zero heavy deps)
- Re-export ActionResult from agent/views.py for backward compatibility
- Update tools/service.py to import from models and registry.views
- Update __init__.py lazy imports to point to new locations
